PMM-10562 Simplified PSMDB cluster creation
